Add files via upload
Improvements to auto-detect based on testing with Nexus Liquid Cooled model. LiquidCooled conf option is now optional
Added more display messages for 2-wire start / stop
Added optional conf parameter to disable checking of outages (monitoring of the utility voltage). This is needed if you have a "smart" transfer switch and the utility voltage is not wired to your controller.