No absolute URLs: Remove leading / from file names.
The web-based UI may have to work when embedded (via proxy) into another web
server, at a subdirectory. That means all URLs within the html that is served
need to be relative, not absolute; that includes URLs served from JS and CSS.
Fixed a few instances of URLs that begin with "/images/...".